Added Firebase, App Icon
Beginning to add Firebase to this app.  It will use the same Firebase project I am using for Orlando Walking Tours, with the intent of this being sort of an administration type app.  Also included is the App Icon I made before, and I am adding the Assets folder to the .gitignore.